+++ asf_search/export/jsonlite2.py
@@ -7,7 +7,11 @@ from .jsonlite import JSONLiteStreamArray
 
 def results_to_jsonlite2(results):
     ASF_LOGGER.info('started translating results to jsonlite2 format')
-
+    
+    if len(results) == 0:
+            yield from json.JSONEncoder(indent=2, sort_keys=True).iterencode({'results': []})
+            return
+        
     if not inspect.isgeneratorfunction(results) and not isinstance(results, GeneratorType):
         results = [results]
     
@@ -54,12 +58,16 @@ class JSONLite2StreamArray(JSONLiteStreamArray):
             'pge': p['pgeVersion']
         }
 
-        if 'temporalBaseline' in p.keys() or 'perpendicularBaseline' in p.keys():
+        if 'temporalBaseline' in p.keys():
             result['tb'] = p['temporalBaseline']
+        if 'perpendicularBaseline' in p.keys():
             result['pb'] = p['perpendicularBaseline']
 
         if p.get('burst') is not None: # is a burst product
             result['s1b'] = p['burst']
+
+        if p.get('opera') is not None:
+            result['s1o'] = p['opera']
         
         return result